Implement function - hash power

+++ contracts/interface/ICandidateHub.sol
@@ -2,5 +2,5 @@ pragma solidity 0.6.12;
 
 interface ICandidateHub {
   function canDelegate(address agent) external view returns(bool);
-  function jailValidator(address operateAddress, uint round, int256 fine) external;
+  function jailValidator(address operateAddress, uint256 round, int256 fine) external;
 }
